# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Revolut-inspired reference fintech platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments/transfers
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, failed, cancelled]
      responses:
        "200":
          description: Payments list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentPage"
    post:
      operationId: initiatePayment
      tags: [payments]
      summary: Initiate a payment or transfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiatePaymentRequest"
      responses:
        "202":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{paymentId}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment details
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # FX
...
components:
  schemas:
...
    Payment:
      type: object
      properties:
        paymentId:
          type: string
        userId:
          type: string
        fromAccountId:
          type: string
        toAccountId:
          type: string
        type:
          type: string
          enum: [internal_transfer, bank_transfer, sepa, swift, instant, scheduled]
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed, cancelled]
        reference:
          type: string
        recipientName:
          type: string
        recipientIban:
          type: string
        recipientBic:
          type: string
        scheduledAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [paymentId, userId, fromAccountId, type, amount, currency, status, createdAt]
    PaymentPage:
      type: object
      properties:
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
      required: [payments, total, offset, limit]
    InitiatePaymentRequest:
      type: object
      properties:
        fromAccountId:
          type: string
        type:
          type: string
          enum: [internal_transfer, bank_transfer, sepa, swift, instant]
        amount:
          type: number
        currency:
          type: string
        reference:
          type: string
        recipientName:
          type: string
        recipientIban:
          type: string
        recipientBic:
          type: string
        toAccountId:
          type: string
      required: [fromAccountId, type, amount, currency]
...