# lib/api-zod/src/generated/types/initiatePaymentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Revolut-inspired reference fintech platform API
 * OpenAPI spec version: 0.1.0
 */
import type { InitiatePaymentRequestType } from "./initiatePaymentRequestType";

export interface InitiatePaymentRequest {
  fromAccountId: string;
  type: InitiatePaymentRequestType;
  amount: number;
  currency: string;
  reference?: string;
  recipientName?: string;
  recipientIban?: string;
  recipientBic?: string;
  toAccountId?: string;
}